+++ src/main/java/com/microtomato/hirun/modules/user/entity/po/UserFunc.java
@@ -31,13 +31,13 @@ public class UserFunc extends BaseEntity {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "USER_FUNC_ID", type = IdType.AUTO)
-    private Integer userFuncId;
+    private Long userFuncId;
 
     @TableField("USER_ID")
-    private Integer userId;
+    private Long userId;
 
     @TableField("FUNC_ID")
-    private Integer funcId;
+    private Long funcId;
 
     @TableField("START_DATE")
     private LocalDateTime startDate;
@@ -52,13 +52,13 @@ public class UserFunc extends BaseEntity {
     private String remark;
 
     @TableField("CREATE_USER_ID")
-    private Integer createUserId;
+    private Long createUserId;
 
     @TableField("CREATE_DATE")
     private LocalDateTime createDate;
 
     @TableField("UPDATE_USER_ID")
-    private Integer updateUserId;
+    private Long updateUserId;
 
     @TableField("UPDATE_TIME")
     private LocalDateTime updateTime;
